src/rag_worker/requirements.txt
fastapi
python-dotenv
openai>=1.14.0
llama-index
llama-index-embeddings-openai
llama-index-vector-stores-supabase
llama-index-vector-stores-postgres
pgvector>=0.2.5
llama-index-program-openai
llama-index-postprocessor-cohere_rerank
dataclasses-json
supabase
requests
uvicorn
boto3
debugpy
pyinstrument
arize-phoenix==10.14.0
arize-phoenix-otel
opentelemetry
opentelemetry-sdk
opentelemetry-exporter-otlp
openinference-instrumentation-openai
openinference-instrumentation-llama_index